refactor(schematics): run tests for template files
* No longer skips tests that verify that the template files are properly created as part of the schematic. This is now possible because we have a more clean workaround for the schematic testing (related: https://github.com/bazelbuild/rules_typescript/issues/154)
* Updates all schema.json files to properly match the current implementation of the `build-component` function (consistent with `ng g component`)
* Due to the new workaround for the schematic testing we can also now re-add the `schema.json` files to the collection. The schema files are needed for the schematics when using `ng generate`)
