+++ src/lib/schematics/table/schema.json
@@ -13,7 +13,6 @@
     "project": {
       "type": "string",
       "description": "The name of the project.",
-      "visible": false,
       "$default": {
         "$source": "projectName"
       }
@@ -91,6 +90,11 @@
       "type": "boolean",
       "default": false,
       "description": "Specifies if declaring module exports the component."
+    },
+    "entryComponent": {
+      "type": "boolean",
+      "default": false,
+      "description": "Specifies if the component is an entry component of declaring module."
     }
   },
   "required": []
