+++ src/lib/schematics/test-setup/post-scheduled-tasks.ts
@@ -7,36 +7,9 @@
  */
 
 import {EngineHost, TaskScheduler} from '@angular-devkit/schematics';
-import {SchematicTestRunner, UnitTestTree} from '@angular-devkit/schematics/testing';
-import {join} from 'path';
+import {SchematicTestRunner} from '@angular-devkit/schematics/testing';
 import {from as observableFrom, Observable} from 'rxjs';
 import {concatMap, filter, last} from 'rxjs/operators';
-
-/** Path to the test collection file for the Material schematics */
-export const collectionPath = join(__dirname, '..', 'test-collection.json');
-
-/** Path to the test migration file for the Material update schematics */
-export const migrationCollection = join(__dirname, '..', 'test-migration.json');
-
-/** Create a base app used for testing. */
-export function createTestApp(): UnitTestTree {
-  const baseRunner = new SchematicTestRunner('material-schematics', collectionPath);
-
-  const workspaceTree = baseRunner.runExternalSchematic('@schematics/angular', 'workspace', {
-    name: 'workspace',
-    version: '6.0.0',
-    newProjectRoot: 'projects',
-  });
-
-  return baseRunner.runExternalSchematic('@schematics/angular', 'application', {
-    name: 'material',
-    inlineStyle: false,
-    inlineTemplate: false,
-    routing: false,
-    style: 'scss',
-    skipTests: false,
-  }, workspaceTree);
-}
 
 /**
  * Due to the fact that the Angular devkit does not support running scheduled tasks from a
